fix(studio): Fix for decoding failure on setting long layer names
EOF should not be treated as an IO error in cases where `pb_read` is not
attempting to read more bytes than are available. After this, `pb_read`
will read one more byte to check for the next tag or EOF, which is the
correct time for that check, as we do at the beginning of this function.